# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EquipChain Global Ltd API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts (public)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 9
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostList"

    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blog/posts/admin:
    get:
      operationId: adminListBlogPosts
      tags: [blog]
      summary: List all blog posts (admin, includes drafts)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, scheduled]
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blog/posts/{slug}:
    get:
      operationId: getBlogPostBySlug
      tags: [blog]
      summary: Get a blog post by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blog/posts/{id}:
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blog/posts/{id}/publish:
    patch:
      operationId: toggleBlogPostPublish
      tags: [blog]
      summary: Toggle publish status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /blog/stats:
    get:
      operationId: getBlogStats
      tags: [blog]
      summary: Blog statistics (admin dashboard)
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Contact
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

    get:
      operationId: listContactSubmissions
      tags: [contact]
      summary: List contact submissions (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [unread, read]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contact submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmissionList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contact/{id}:
    get:
      operationId: getContactSubmission
      tags: [contact]
      summary: Get contact submission by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteContactSubmission
      tags: [contact]
      summary: Delete contact submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contact/{id}/read:
    patch:
      operationId: markContactRead
      tags: [contact]
      summary: Mark contact as read/unread (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Quote Requests
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    BlogPost:
      type: object
      required: [id, title, slug, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        category:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        featuredImage:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, scheduled]
        author:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    BlogPostList:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    BlogPostInput:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        tags:
          type: string
        featuredImage:
          type: string
        excerpt:
          type: string
        content:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        status:
          type: string
          enum: [draft, published, scheduled]
        author:
          type: string
        scheduledAt:
          type: string
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        tags:
          type: string
        featuredImage:
          type: string
        excerpt:
          type: string
        content:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        status:
          type: string
          enum: [draft, published, scheduled]
        author:
          type: string
        scheduledAt:
          type: string
    PublishToggleInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [draft, published, scheduled]
    BlogStats:
      type: object
      required: [total, published, drafts, scheduled]
      properties:
        total:
          type: integer
        published:
          type: integer
        drafts:
          type: integer
        scheduled:
          type: integer
    ContactInput:
      type: object
      required: [fullName, email, message]
      properties:
        fullName:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        inquiryType:
          type: string
        message:
          type: string
    ContactSubmission:
      type: object
      required: [id, fullName, email, message, isRead, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        companyName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        inquiryType:
          type: ["string", "null"]
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    ContactSubmissionList:
      type: object
      required: [submissions, total, page, limit]
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/ContactSubmission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ReadStatusInput:
      type: object
      required: [isRead]
      properties:
        isRead:
          type: boolean
...