# packages/api-zod/src/generated/types/blogPostUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EquipChain Global Ltd API
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPostUpdateStatus } from './blogPostUpdateStatus';

export interface BlogPostUpdate {
  title?: string;
  slug?: string;
  category?: string;
  tags?: string;
  featuredImage?: string;
  excerpt?: string;
  content?: string;
  seoTitle?: string;
  seoDescription?: string;
  status?: BlogPostUpdateStatus;
  author?: string;
  scheduledAt?: string;
}