# packages/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EquipChain Global Ltd API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  fullName: string;
  companyName?: string;
  email: string;
  phone?: string;
  inquiryType?: string;
  message: string;
}