# packages/api-zod/src/generated/types/contactSubmissionList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EquipChain Global Ltd API
 * OpenAPI spec version: 0.1.0
 */
import type { ContactSubmission } from './contactSubmission';

export interface ContactSubmissionList {
  submissions: ContactSubmission[];
  total: number;
  page: number;
  limit: number;
}